Update preCompileScript.cmd
This code sets the SDK_RELEASE_YEAR and SDK_RELEASE_MONTH values to 2022 and 04, respectively. It also adds a call to vcvarsamd64_x86.bat to set up the Visual Studio environment.

It also uses delayed expansion to set the values of SDK_RELEASE_YEAR and SDK_RELEASE_MONTH based on the contents of the BuildVersion.txt file. The script will search for BuildVersion.txt in several locations if it's not found in the current directory.