Update page.js
This is a set of JavaScript functions that are used for expanding and collapsing content in a web page. The toggleChildren function toggles the visibility of the child elements of the given element, while the showChildren function shows the child elements. The toggleNext function toggles the visibility of the next sibling element of the given element. The toggleId function toggles the visibility of the child elements of the element with the given ID. The jumpToId function scrolls the page to the element with the given ID and shows its child elements. The jumpToCall function scrolls the page to the element with the given ID. The jumpToTop function scrolls the page to the top.