Update Constants.cs
This code defines a C# namespace called XboxLiveTrace that contains two types: a class called Constants, and an enum called CSVValueIndex.

Constants defines several constant string and numerical values that are used elsewhere in the application. These values are used to define things like time periods, allowed call frequencies, and other metrics.

CSVValueIndex is an enumeration of integer values that are used to specify the index of various fields in a CSV file. These values are used to parse the data in the CSV file, and they correspond to specific fields in the CSV file format.